Update mixed precision

+++ models/CTCModel.py
@@ -20,7 +20,7 @@ def create_ctc_model(base_model, num_classes, speech_conf, streaming_size=None, 
 
     # Fully connected layer
     outputs = tf.keras.layers.TimeDistributed(
-        tf.keras.layers.Dense(units=num_classes, activation="linear",
+        tf.keras.layers.Dense(units=num_classes, activation="linear", dtype=tf.float32,
                               use_bias=True), name="fully_connected")(outputs)
 
     model = tf.keras.Model(inputs=features, outputs=outputs, name=name)
